+++ src/WeCantSpell.Hunspell/Infrastructure/StringBuilderEx.cs
@@ -171,6 +171,36 @@ namespace WeCantSpell.Hunspell.Infrastructure
             }
         }
 
+        public static void WriteChars(this StringBuilder @this, StringSlice text, int destinationIndex)
+        {
+            if (destinationIndex >= @this.Length)
+            {
+                if (destinationIndex > @this.Length)
+                {
+                    var characterGap = destinationIndex - @this.Length;
+                    @this.Append('\0', characterGap);
+                }
+
+                @this.Append(text);
+            }
+            else
+            {
+                var writeUntilIndex = destinationIndex + text.Length;
+                if (writeUntilIndex <= @this.Length)
+                {
+                    for (var i = 0; i < text.Length; i++)
+                    {
+                        @this[destinationIndex + i] = text[i];
+                    }
+                }
+                else
+                {
+                    @this.Remove(destinationIndex, @text.Length - destinationIndex);
+                    @this.Append(text);
+                }
+            }
+        }
+
         public static void WriteChars(this StringBuilder @this, int sourceIndex, string text, int destinationIndex)
         {
             if (destinationIndex >= @this.Length)
@@ -252,6 +282,11 @@ namespace WeCantSpell.Hunspell.Infrastructure
             return builder.ToString(0, index)
                 + value.ToString()
                 + builder.ToString(index, builder.Length - index);
+        }
+
+        public static void AppendSlice(this StringBuilder builder, StringSlice slice)
+        {
+            builder.Append(slice.Text, slice.Offset, slice.Length);
         }
     }
 }