+++ changelog.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - Adds phonetic entries to the replacement list
 - Restricts compound replacement to using "middle" entries
 - Reduce number of strange ngram suggestions
+- Prefer suggestions for word pairs listed in dictionary
 
 ## 2.0.1 - 2018-07-22
 ### Fixed